+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
  - Fix timestamp argument having no affect in value_set()
  - Revert changes to PCF8574 Output ([#1175](https://github.com/kizniche/mycodo/issues/1175))
+ - Fix Edge Detection Input storing measurements ([#1171](https://github.com/kizniche/mycodo/issues/1171))
 
 ### Features